requirements.txt
asgiref==3.2.7
autopep8==1.5.1
beautifulsoup4==4.9.0
camelot-py==0.7.3
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreschema==0.0.4
covid==2.2.0
dataclasses==0.7
dateutils==0.6.7
Deprecated==1.2.9
Django==3.0.5
django-chartjs==2.0.0
django-filter==2.2.0
django-tables2==2.3.0
djangorestframework==3.11.0
drf-yasg==1.17.1
et-xmlfile==1.0.1
facebook-sdk==3.1.0
ghostscript==0.6
gitdb==4.0.4
GitPython==3.1.1
idna==2.9
inflection==0.4.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.18.2
opencv-python==4.2.0.34
openpyxl==3.0.3
packaging==20.3
pandas==1.0.3
pdfminer.six==20200402
Pillow==7.1.2
pycodestyle==2.5.0
pycryptodome==3.9.7
pydantic==1.4
PyGithub==1.47
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
selenium==3.141.0
six==1.14.0
smmap==3.0.2
sortedcontainers==2.1.0
soupsieve==2.0
SQLAlchemy==1.3.15
sqlparse==0.3.1
tk==0.1.0
uritemplate==3.0.1
urllib3==1.25.8
wrapt==1.12.1
